Quickly browse to important package webpages

These functions take you to various webpages associated with a package and return the target URL invisibly. Some URLs are formed from first principles and there is no guarantee there will be content at the destination.

browse_github(package = NULL)

browse_github_issues(package = NULL, number = NULL)

browse_github_pulls(package = NULL, number = NULL)

browse_travis(package = NULL)

browse_cran(package = NULL)

Arguments

package
Name of package; leave as NULL to use current package

number
For GitHub issues and pull requests. Can be a number or "new".

Details

browse_github(): Looks for a GitHub URL in the URL field of DESCRIPTION.

browse_github_issues(): Visits the GitHub Issues index or one specific issue.

browse_github_pulls(): Visits the GitHub Pull Request index or one specific pull request.

browse_travis(): Visits the package's page on Travis CI.

browse_cran(): Visits the package on CRAN, via the canonical URL.

Examples

browse_github("gh")

#> ● Open url https://github.com/r-lib/gh#readme

browse_github_issues("backports")

#> ● Open url https://github.com/r-lib/backports/issues/

browse_github_issues("backports", 1)

#> ● Open url https://github.com/r-lib/backports/issues/1

browse_github_pulls("rprojroot")

#> ● Open url https://github.com/krlmlr/rprojroot/pulls/

browse_github_pulls("rprojroot", 3)

#> ● Open url https://github.com/krlmlr/rprojroot/pull/3

browse_travis("usethis")

#> ● Open url https://travis-ci.org/r-lib/usethis

browse_cran("MASS")

#> ● Open url https://cran.r-project.org/package=MASS