Continuous integration setup and badges

Sets up continuous integration (CI) services for an R package that is developed on GitHub. CI services can run R CMD check automatically on various platforms, triggered by each push or pull request. These functions

Add service-specific configuration files and add them to .Rbuildignore.

Activate a service or give the user a detailed prompt.

Provide the markdown to insert a badge into README.

use_travis(browse = interactive(), ext = c("org", "com"))

use_appveyor(browse = interactive())

use_gitlab_ci()

Arguments

browse
Open a browser window to enable automatic builds for the package.

ext
which travis website to use. default to "org"for https://travis-ci.org. Change to "com" for https://travis-ci.com.

use_travis()

Adds a basic .travis.yml to the top-level directory of a package. This is a configuration file for the Travis CI continuous integration service.

use_appveyor()

Adds a basic appveyor.yml to the top-level directory of a package. This is a configuration file for the AppVeyor continuous integration service for Windows.

use_gitlab_ci()

Adds a basic .gitlab-ci.yml to the top-level directory of a package. This is a configuration file for the GitLab CI/CD continuous integration service for GitLab.