edit {usethis}	R Documentation

Open configuration files

Description

edit_r_profile() opens .Rprofile

edit_r_environ() opens .Renviron

edit_r_makevars() opens .R/Makevars

edit_git_config() opens .gitconfig or .git/config

edit_git_ignore() opens .gitignore

edit_rstudio_snippets(type) opens .R/snippets/{type}.snippets

Usage

edit_r_profile(scope = c("user", "project"))

edit_r_environ(scope = c("user", "project"))

edit_r_buildignore(scope = c("user", "project"))

edit_r_makevars(scope = c("user", "project"))

edit_rstudio_snippets(type = "R")

edit_git_config(scope = c("user", "project"))

edit_git_ignore(scope = c("user", "project"))

Arguments

scope
Edit globally for the current user, or locally for the current project

type
Snippet type. One of: "R", "markdown", "C_Cpp", "Tex", "Javascript", "HTML", "SQL"

Details

The edit_r_*() and edit_rstudio_*() functions consult R's notion of user's home directory. The edit_git_*() functions – and usethis in general – inherit home directory behaviour from the fs package, which differs from R itself on Windows. The fs default is more conventional in terms of the location of user-level Git config files. See fs::path_home() for more details.

Value

Path to the file, invisibly.

[Package usethis version 1.5.0 Index]