Use MIT, GPL-3, or Apache 2.0 license for your package

Adds the necessary infrastructure to declare your package as licensed with one of four popular open source license:

CC0: dedicated to public domain. Appropriate for data packages.

MIT: simple and permissive.

Apache 2.0: provides patent protection.

GPL v3: requires sharing of improvements.

See https://choosealicense.com for more details and other options.

use_mit_license(name = find_name())

use_gpl3_license(name = find_name())

use_apl2_license(name = find_name())

use_cc0_license(name = find_name())

Arguments

name
Name of the copyright holder or holders. Separate multiple individuals with ;. You can supply a global default with options(usethis.full_name = "My name").

Details

CRAN does not allow you to include copies of standard licenses in your package, so these functions save the license as LICENSE.md and add it to .Rbuildignore.