Helpers for the tidyverse

These helpers follow tidyverse conventions which are generally a little stricter than the defaults, reflecting the need for greater rigor in commonly used packages.

use_tidy_ci(browse = interactive())

use_tidy_description()

use_tidy_versions(overwrite = FALSE)

use_tidy_eval()

use_tidy_contributing()

use_tidy_issue_template()

use_tidy_support()

use_tidy_coc()

use_tidy_github()

use_tidy_style(strict = TRUE)

Arguments

browse
Open a browser window to enable automatic builds for the package.

overwrite
By default (FALSE), only dependencies without version specifications will be modified. Set to TRUE to modify all dependencies.

strict
Boolean indicating whether or not a strict version of styling should be applied. See styler::tidyverse_style() for details.

Details

use_tidy_ci(): sets up travis and codecov, ensuring that the package works on all versions of R starting at 3.1.

use_tidy_description(): puts fields in standard order and alphabetises dependencies.

use_tidy_eval(): imports a standard set of helpers to facilitate programming with the tidy eval toolkit.

use_tidy_style(): styles source code according to the tidyverse style guide. This function will overwrite files! See below for usage advice.

use_tidy_versions(): pins all dependencies to require at least the currently installed version.

use_tidy_contributing(): adds standard tidyverse contributing guidelines.

use_tidy_issue_template(): adds a standard tidyverse issue template.

use_tidy_support(): adds a standard description of support resources for the tidyverse.

use_tidy_coc(): equivalent to use_code_of_conduct(), but puts the document in a .github/ subdirectory.

use_tidy_github(): convenience wrapper that calls use_tidy_contributing(), use_tidy_issue_template(), use_tidy_support(), use_tidy_coc().

<code>use_tidy_style()</code>

Uses the styler package package to style all code in a package, project, or directory, according to the tidyverse style guide. Warning: This function will overwrite files! It is strongly suggested to only style files that are under version control or to first create a backup copy.

Invisibly returns a data frame with one row per file, that indicates whether styling caused a change.