use_github_actions {usethis}	R Documentation

Defunct GitHub Actions workflows

Description

use_github_actions() is deprecated because it was just an alias for use_github_action_check_release().

use_github_action_check_full() is overkill for most packages and is not recommended.

use_github_action_check_release(), use_github_action_check_standard(), and use_github_action_pr_commands() are deprecated in favor of use_github_action(), which can now suggest specific workflows to use.

Usage


use_github_actions()

use_github_action_check_release(
  save_as = "R-CMD-check.yaml",
  ref = NULL,
  ignore = TRUE,
  open = FALSE
)

use_github_action_check_standard(
  save_as = "R-CMD-check.yaml",
  ref = NULL,
  ignore = TRUE,
  open = FALSE
)

use_github_action_pr_commands(
  save_as = "pr-commands.yaml",
  ref = NULL,
  ignore = TRUE,
  open = FALSE
)

use_github_action_check_full(
  save_as = "R-CMD-check.yaml",
  ignore = TRUE,
  open = FALSE,
  repo_spec = NULL
)

[Package usethis version 2.2.3 Index]