use_lifecycle {usethis}	R Documentation

Use lifecycle badges

Description

This helper adds lifecycle as a dependency, copies the lifecycle badges into the man/figures folder, and reminds you how to use the badge syntax.

Learn more at https://lifecycle.r-lib.org/articles/communicate.html

Usage

use_lifecycle()

See Also

use_lifecycle_badge() to signal the lifecycle stage of your package as whole

[Package usethis version 2.0.1 Index]