use_readme_rmd {usethis}	R Documentation

Create README files

Description

Creates skeleton README files with sections for

a high-level description of the package and its goals

R code to install from GitHub, if GitHub usage detected

a basic example

Use Rmd if you want a rich intermingling of code and output. Use md for a basic README. README.Rmd will be automatically added to .Rbuildignore. The resulting README is populated with default YAML frontmatter and R fenced code blocks (md) or chunks (Rmd).

If you use Rmd, you'll still need to render it regularly, to keep README.md up-to-date. devtools::build_readme() is handy for this. You could also use GitHub Actions to re-render README.Rmd every time you push. An example workflow can be found in the examples/ directory here: https://github.com/r-lib/actions/.

Usage

use_readme_rmd(open = rlang::is_interactive())

use_readme_md(open = rlang::is_interactive())

Arguments

open
Open the newly created file for editing? Happens in RStudio, if applicable, or via utils::file.edit() otherwise.

See Also

The important files section of R Packages.

Examples

## Not run: 
use_readme_rmd()
use_readme_md()

## End(Not run)

[Package usethis version 2.1.5 Index]