use_release_issue {usethis}	R Documentation

Create a release checklist in a GitHub issue

Description

When preparing to release a package there are quite a few steps that need to be performed, and some of the steps can take multiple hours. This function creates an issue checklist so that you can keep track of where you are in the process, and feel a sense of satisfaction as you progress. It also helps watchers of your package stay informed about where you are in the process.

Usage

use_release_issue(version = NULL)

Arguments

version
Optional version number for release. If unspecified, you can make an interactive choice.

Examples

## Not run: 
use_release_issue("2.0.0")

## End(Not run)

[Package usethis version 2.0.1 Index]