Add an RMarkdown Template

Adds files and directories necessary to add a custom rmarkdown template to RStudio. It creates:

inst/rmarkdown/templates/{{template_dir}}. Main directory.

skeleton/skeleton.Rmd. Your template Rmd file.

template.yml with basic information filled in.

use_rmarkdown_template(template_name = "Template Name",
  template_dir = slug(template_name, ""),
  template_description = "A description of the template",
  template_create_dir = FALSE)

Arguments

template_name
The name as printed in the template menu.

template_dir
Name of the directory the template will live in within inst/rmarkdown/templates.

template_description
Sets the value of description in template.yml.

template_create_dir
Sets the value of create_dir in template.yml.

Examples

# NOT RUN {
use_rmarkdown_template()
# }