use_rstudio_preferences {usethis}	R Documentation

Set global RStudio preferences

Description

This function allows you to set global RStudio preferences, achieving the same effect programmatically as clicking buttons in RStudio's Global Options. You can find a list of configurable properties at https://docs.posit.co/ide/server-pro/reference/session_user_settings.html.

Usage


use_rstudio_preferences(...)

Arguments

...
<dynamic-dots> Property-value pairs.

Value

A named list of the previous values, invisibly.

[Package usethis version 2.2.3 Index]