use_testthat {usethis}	R Documentation

Sets up overall testing infrastructure

Description

Creates tests/testthat/, tests/testthat.R, and adds the testthat package to the Suggests field. Learn more in https://r-pkgs.org/tests.html

Usage

use_testthat(edition = NULL, parallel = FALSE)

Arguments

edition
testthat edition to use. Defaults to the latest edition, i.e. the major version number of the currently installed testthat.

parallel
Should tests be run in parallel? This feature appeared in testthat 3.0.0; see https://testthat.r-lib.org/articles/parallel.html for details and caveats.

See Also

use_test() to create individual test files

Examples

## Not run: 
use_testthat()

use_test()

use_test("something-management")

## End(Not run)

[Package usethis version 2.1.5 Index]