use_tidy_labels {usethis}	R Documentation

Defunct tidyverse functions

Description

use_tidy_labels() has been replaced by use_tidy_github_labels().

use_tidy_ci() has been replaced by use_tidy_github_actions().

use_tidy_eval() is defunct because there's no longer a need to systematically import and re-export a large number of functions in order to use tidy evaluation. Instead, use use_import_from() to tactically import functions as you need them.

Usage


use_tidy_labels()

use_tidy_ci(...)

use_tidy_eval()

[Package usethis version 2.2.3 Index]