Setup deployment for Travis CI

Creates a public-private key pair, adds the public key to the GitHub repository via github_add_key(), and stores the private key as an encrypted environment variable in Travis CI via travis_set_var(), possibly in a different repository. The tic companion package contains facilities for installing such a key during a Travis CI build.

use_travis_deploy(path = ".", info = travis:::github_info(path),
  repo = travis:::github_repo(info = info), travis_repo = repo)

Arguments

path
[string]
The path to a GitHub-enabled Git repository (or a subdirectory thereof).

info
[list]
GitHub information for the repository, by default obtained through github_info().

repo
[string|numeric]
The GitHub repo slug, by default obtained through github_repo(). Alternatively, the Travis CI repo ID, e.g. obtained through travis_repo_id().

travis_repo
[string]
The Travis CI repository to add the private key to, default: repo (the GitHub repo to which the public deploy key is added).