use_upkeep_issue {usethis}	R Documentation

Create an upkeep checklist in a GitHub issue

Description

This opens an issue in your package repository with a checklist of tasks for regular maintenance of your package. This is a fairly opinionated list of tasks but we believe taking care of them will generally make your package better, easier to maintain, and more enjoyable for your users. Some of the tasks are meant to be performed only once (and once completed shouldn't show up in subsequent lists), and some should be reviewed periodically. The tidyverse team uses a similar function use_tidy_upkeep_issue() for our annual package Spring Cleaning.

Usage


use_upkeep_issue(year = NULL)

Arguments

year
Year you are performing the upkeep, used in the issue title. Defaults to current year

Examples


## Not run: 
use_upkeep_issue(2023)

## End(Not run)

[Package usethis version 2.2.3 Index]