use_vignette {usethis}	R Documentation

Create a vignette

Description

Performs general set up for vignettes and initializes a new individual vignette:

Adds needed packages to DESCRIPTION

Adds inst/doc to .gitignore so built vignettes aren't tracked

Creates a new draft .Rmd vignette in vignettes/ and, if possible, opens it for editing

Usage

use_vignette(name)

Arguments

name
Base for file name to use for new vignette. Should consist only of numbers, letters, _ and -. I recommend using lower case.

See Also

The vignettes chapter of R Packages.

Examples

## Not run: 
use_vignette("how-to-do-stuff")

## End(Not run)

[Package usethis version 1.4.0 Index]