usethis: Automate Package and Project Setup

Automate package and project setup tasks that are otherwise performed manually. This includes setting up unit testing, test coverage, continuous integration, Git, 'GitHub', licenses, 'Rcpp', 'RStudio' projects, and more.

See also

Author

Maintainer: Jennifer Bryan jenny@rstudio.com (ORCID)

Authors:

Hadley Wickham hadley@rstudio.com (ORCID)

Malcolm Barrett malcolmbarrett@gmail.com (ORCID)

Other contributors:

RStudio [copyright holder, funder]