badges {usethis}	R Documentation

README badges

Description

These helpers produce the markdown text you need in your README to include badges that report information, such as the CRAN version or test coverage, and link out to relevant external resources. To add badges automatically ensure your badge block starts with a line containing only <!-- badges: start --> and ends with a line containing only <!-- badges: end -->.

Usage

use_badge(badge_name, href, src)

use_cran_badge()

use_bioc_badge()

use_lifecycle_badge(stage)

use_binder_badge(ref = git_default_branch(), urlpath = NULL)

Arguments

badge_name
Badge name. Used in error message and alt text

href, src
Badge link and image src

stage
Stage of the package lifecycle. One of "experimental", "stable", "superseded", or "deprecated".

ref
A Git branch, tag, or SHA

urlpath
An optional urlpath component to add to the link, e.g. "rstudio" to open an RStudio IDE instead of a Jupyter notebook. See the binder documentation for additional examples.

Details

use_badge(): a general helper used in all badge functions

use_bioc_badge(): badge indicates BioConductor build status

use_cran_badge(): badge indicates what version of your package is available on CRAN, powered by https://www.r-pkg.org

use_lifecycle_badge(): badge declares the developmental stage of a package according to https://lifecycle.r-lib.org/articles/stages.html.

use_binder_badge(): badge indicates that your repository can be launched in an executable environment on https://mybinder.org/

See Also

Functions that configure continuous integration, such as use_github_actions(), also create badges.

Examples

## Not run: 
use_cran_badge()
use_lifecycle_badge("stable")

## End(Not run)

[Package usethis version 2.1.5 Index]