README badges

These helpers produce the markdown text you need in your README to include badges that report information, such as the CRAN version or test coverage, and link out to relevant external resources.

use_badge(badge_name, href, src)

use_cran_badge()

use_bioc_badge()

use_depsy_badge()

use_lifecycle_badge(stage)

use_binder_badge()

Arguments

badge_name
Badge name. Used in error message and alt text

href, src
Badge link and image src

stage
Stage of the package lifecycle

Details

use_badge(): a general helper used in all badge functions

use_bioc_badge(): badge indicates BioConductor buildstatus

use_cran_badge(): badge indicates what version of your package is available on CRAN, powered by https://www.r-pkg.org

use_depsy_badge(): badge shows the "percentile overall impact" of the project, powered by http://depsy.org, which only indexes projects that are on CRAN

use_lifecycle_badge(): badge declares the developmental stage of a package, according to https://www.tidyverse.org/lifecycle/:

Experimental

Maturing

Stable

Retired

Archived

Dormant

Questioning

use_binder_badge(): badge indicates that your repository can be launched in an executable environment on https://mybinder.org/

See also

The functions that set up continuous integration services also create badges.

Examples

# NOT RUN {
use_cran_badge()
use_lifecycle_badge("stable")
# }