Create a GitHub personal access token

Opens a browser window to the GitHub page where you can generate a Personal Access Token. Make sure you have signed up for a free GitHub.com account and that you are signed in. Click "Generate token" after you have verified the scopes. Copy the token right away! You probably want to store it in .Renviron as the GITHUB_PAT environment variable. edit_r_environ() can help you do that. Use gh::gh_whoami() to get information on an existing token.

browse_github_pat(scopes = c("repo", "gist"), description = "R:GITHUB_PAT",
  host = "https://github.com")

Arguments

scopes
Character vector of token permissions. These are just defaults that will be pre-selected in the web form. You can select the final values on the GitHub page. Read more about GitHub API scopes at https://developer.github.com/apps/building-oauth-apps/scopes-for-oauth-apps/.

description
Short description or nickname for the token. It helps you distinguish various tokens on GitHub.

host
GitHub API host to use. Override with the endpoint-root for your GitHub enterprise instance, for example, "https://github.hostname.com/api/v3"

Examples

# NOT RUN {
browse_github_pat()
## COPY THE PAT!!!
## almost certainly to be followed by ...
edit_r_environ()
## which helps you store the PAT as an env var
# }