ci {usethis}	R Documentation

Continuous integration setup and badges

Description

Sets up continuous integration (CI) services for an R package that is developed on GitHub. CI services can run R CMD check automatically on various platforms, triggered by each push or pull request. These functions

Add service-specific configuration files and add them to .Rbuildignore.

Activate a service or give the user a detailed prompt.

Provide the markdown to insert a badge into README.

Usage

use_travis(browse = interactive())

use_coverage(type = c("codecov", "coveralls"))

use_appveyor(browse = interactive())

Arguments

browse
Open a browser window to enable automatic builds for the package.

type
Which web service to use for test reporting. Currently supports Codecov and Coveralls.

use_travis()

Adds a basic .travis.yml to the top-level directory of a package. This is a configuration file for the Travis CI continuous integration service.

use_coverage()

Adds test coverage reports to a package that is already using Travis CI.

use_appveyor()

Adds a basic appveyor.yml to the top-level directory of a package. This is a configuration file for the AppVeyor continuous integration service for Windows.

[Package usethis version 1.4.0 Index]