ci {usethis}	R Documentation

Continuous integration setup and badges

Description

Some of these functions are now soft-deprecated since the tidyverse team has started using GitHub Actions (GHA) for continuous integration (CI). See use_github_actions() for help configuring GHA. GHA functionality in usethis is actively maintained and exercised, which is no longer true for Travis-CI or AppVeyor.

Sets up third-party continuous integration (CI) services for an R package that is developed on GitHub or, perhaps, GitLab. These functions

Add service-specific configuration files and add them to .Rbuildignore.

Activate a service or give the user a detailed prompt.

Provide the markdown to insert a badge into README.

Usage

use_travis(browse = rlang::is_interactive(), ext = c("com", "org"))

use_travis_badge(ext = c("com", "org"), repo_spec = NULL)

use_appveyor(browse = rlang::is_interactive())

use_appveyor_badge(repo_spec = NULL)

use_gitlab_ci()

use_circleci(browse = rlang::is_interactive(), image = "rocker/verse:latest")

use_circleci_badge(repo_spec = NULL)

Arguments

browse
Open a browser window to enable automatic builds for the package.

ext
Which travis website to use. Defaults to "com" for https://www.travis-ci.com/. Change to "org" for https://travis-ci.org.

repo_spec
Optional GitHub repo specification in this form: owner/repo. This can usually be inferred from the GitHub remotes of active project.

image
The Docker image to use for build. Must be available on DockerHub. The rocker/verse image includes TeXLive, pandoc, and the tidyverse packages. For a minimal image, try rocker/r-ver. To specify a version of R, change the tag from latest to the version you want, e.g. rocker/r-ver:3.5.3.

use_travis()

Adds a basic .travis.yml to the top-level directory of a package. This is a configuration file for the Travis CI continuous integration service.

use_travis_badge()

Only adds the Travis CI badge. Use for a project where Travis is already configured.

use_appveyor()

Adds a basic appveyor.yml to the top-level directory of a package. This is a configuration file for the AppVeyor continuous integration service for Windows.

use_appveyor_badge()

Only adds the AppVeyor badge. Use for a project where AppVeyor is already configured.

use_gitlab_ci()

Adds a basic .gitlab-ci.yml to the top-level directory of a package. This is a configuration file for the GitLab CI/CD continuous integration service.

use_circleci()

Adds a basic .circleci/config.yml to the top-level directory of a package. This is a configuration file for the CircleCI continuous integration service.

use_circleci_badge()

Only adds the Circle CI badge. Use for a project where Circle CI is already configured.

[Package usethis version 2.1.5 Index]