Continuous integration setup and badges

Sets up continuous integration (CI) services for an R package that is developed on GitHub. CI services can run R CMD check automatically on various platforms, triggered by each push or pull request. These functions

Add service-specific configuration files and add them to .Rbuildignore.

Activate a service or give the user a detailed prompt.

Provide the markdown to insert a badge into README.

use_travis(browse = interactive(), ext = c("org", "com"))

use_travis_badge(ext = "org")

use_appveyor(browse = interactive())

use_appveyor_badge()

use_gitlab_ci()

use_circleci(browse = interactive(), image = "rocker/verse:latest")

use_circleci_badge()

Arguments

browse
Open a browser window to enable automatic builds for the package.

ext
which travis website to use. default to "org"for https://travis-ci.org. Change to "com" for https://travis-ci.com.

image
The Docker image to use for build. Must be available on DockerHub. The rocker/verse image includes TeX Live, pandoc, and the tidyverse packages. For a minimal image, try rocker/r-ver. To specify a version of R, change the tag from latest to the version you want, e.g. rocker/r-ver:3.5.3.

use_travis()

Adds a basic .travis.yml to the top-level directory of a package. This is a configuration file for the Travis CI continuous integration service.

use_travis_badge()

Only adds the Travis CI badge. Use for a project where Travis is already configured.

use_appveyor()

Adds a basic appveyor.yml to the top-level directory of a package. This is a configuration file for the AppVeyor continuous integration service for Windows.

use_appveyor_badge()

Only adds the AppVeyor badge. Use for a project where AppVeyor is already configured.

use_gitlab_ci()

Adds a basic .gitlab-ci.yml to the top-level directory of a package. This is a configuration file for the GitLab CI/CD continuous integration service for GitLab.

use_circleci()

Adds a basic .circleci/config.yml to the top-level directory of a package. This is a configuration file for the CircleCI continuous integration service.

use_circleci_badge()

Only adds the Circle CI badge. Use for a project where Circle CI is already configured.