git_credentials {usethis}	R Documentation

Produce or register credentials for git2r

Description

In usethis v2.0.0, usethis switched from git2r to gert (+ credentials) for all Git operations. This pair of packages (gert + credentials) is designed to discover and use the same credentials as command line Git. As a result, a great deal of credential-handling assistance has been removed from usethis, primarily around SSH keys.

If you have credential problems, focus your troubleshooting on getting the credentials package to find your credentials. The introductory vignette is a good place to start.

If you use the HTTPS protocol (which we recommend), a GitHub personal access token will satisfy all auth needs, for both Git and the GitHub API, and is therefore the easiest approach to get working. See gh_token_help() for more.

Usage

git_credentials(protocol = deprecated(), auth_token = deprecated())

use_git_credentials(credentials = deprecated())

Arguments

protocol
Deprecated.

auth_token
Deprecated.

credentials
Deprecated.

Value

These functions raise a warning and return an invisible NULL.

[Package usethis version 2.1.5 Index]