github_actions {usethis}	R Documentation

GitHub Actions setup and badges

Description

Sets up continuous integration (CI) for an R package that is developed on GitHub using GitHub Actions. CI can be used to trigger various operations for each push or pull request, such as:

Run R CMD check on various operating systems and R versions

Build and deploy a pkgdown site

Determine test coverage

This family of functions

Adds the necessary configuration files and lists them in .Rbuildignore.

Provides the markdown to insert a badge into your README.

Usage

use_github_actions()

use_github_actions_badge(name = "R-CMD-check", repo_spec = NULL)

Arguments

name
Specifies the workflow whose status the badge will report. This is the name keyword that appears in the workflow .yaml file.

repo_spec
Optional GitHub repo specification in this form: owner/repo. This can usually be inferred from the GitHub remotes of active project.

use_github_actions()

Configures a basic R CMD check workflow on GitHub Actions by adding a standard R-CMD-check.yaml file to the .github/workflows directory of the active project.

use_github_actions_badge()

Generates a GitHub Actions badge and that's all. It does not configure a workflow.

See Also

use_github_action() sets up specific, individual actions, e.g. test coverage or pkgdown build and deploy.

use_tidy_github_actions() sets up the standard GitHub Actions used for tidyverse packages.

[Package usethis version 2.0.1 Index]