github_actions {usethis}	R Documentation

GitHub Actions setup and badges

Description

Sets up continuous integration (CI) for an R package that is developed on GitHub using GitHub Actions. These functions

Add the necessary configuration files and place them in .Rbuildignore.

Provide the markdown to insert a badge into your README

Usage

use_github_actions()

use_github_actions_badge(name = "R-CMD-check")

Arguments

name
The name to give to the GitHub Actions workflow.

use_github_actions()

Adds a basic R-CMD-check.yaml file to the .github/workflows directory of a package. This is a configuration file for the GitHub Actions service.

use_github_actions_badge()

Only adds the GitHub Actions badge. Use for a project where GitHub Actions is already configured.

See Also

use_github_action() for setting up a specific action.

[Package usethis version 1.6.1 Index]