License a package

Adds the necessary infrastructure to declare your package as licensed with one of these popular open source licenses:

CC0: dedicated to public domain. Appropriate for data packages.

MIT: simple and permissive.

Apache 2.0: provides patent protection.

GPL v3: requires sharing of improvements.

AGPL v3: requires sharing of improvements.

LGPL v3: requires sharing of improvements.

CCBY 4.0: Free to share and adapt, must give appropriate credit. Appropriate for data packages.

See https://choosealicense.com for more details and other options.

use_mit_license(name = find_name())

use_gpl3_license(name = find_name())

use_agpl3_license(name = find_name())

use_lgpl_license(name = find_name())

use_apl2_license(name = find_name())

use_cc0_license(name = find_name())

use_ccby_license(name = find_name())

Arguments

name
Name of the copyright holder or holders. Separate multiple individuals with ;. You can supply a global default with options(usethis.full_name = "My name").

Details

CRAN does not allow you to include copies of standard licenses in your package, so these functions save the license as LICENSE.md and add it to .Rbuildignore.

See also