proj_activate {usethis}	R Documentation

Activate a project

Description

Activates a project in usethis, R session, and (if relevant) RStudio senses. If you are in RStudio, this will open a new RStudio session. If not, it will change the working directory and active project.

Usage


proj_activate(path)

Arguments

path
Project directory

Value

Single logical value indicating if current session is modified.

[Package usethis version 2.2.3 Index]