Report working directory and usethis/RStudio project

proj_sitrep() reports

current working directory

the active usethis project

the active RStudio Project

Call this function if things seem weird and you're not sure what's wrong or how to fix it. Usually, all three of these should coincide (or be unset) and proj_sitrep() provides suggested commands for getting back to this happy state.

proj_sitrep()

Value

A named list, with S3 class sitrep (for printing purposes), reporting current working directory, active usethis project, and active RStudio Project

See also

Examples

proj_sitrep()

#>       working_directory:     '/tmp/RtmpPd60cG/file425b286beae/dev/reference'
#>     active_usethis_proj:     NULL
#>     active_rstudio_proj:     NULL
#> 
#> ● There is currently no active usethis project.
#>   usethis attempts to activate a project upon first need.
#>   Call `proj_get()` to explicitly initiate project activation.