use_code_of_conduct {usethis}	R Documentation

Add a code of conduct

Description

Adds a CODE_OF_CONDUCT.md file to the active project and lists in .Rbuildignore, in the case of a package. The goal of a code of conduct is to foster an environment of inclusiveness, and to explicitly discourage inappropriate behaviour. The template comes from http://contributor-covenant.org, version 1: http://contributor-covenant.org/version/1/0/0/.

Usage

use_code_of_conduct(path = NULL)

Arguments

path
Path of the directory to put CODE_OF_CONDUCT.md in, relative to the active project. Passed along to use_directory(). Default is to locate at top-level, but .github/ is also common.

[Package usethis version 1.5.0 Index]