Create or modify a DESCRIPTION file

usethis consults the following sources, in this order, to set DESCRIPTION fields:

fields argument of create_package() or use_description().

getOption("usethis.description") or getOption("devtools.desc"). The devtools option is consulted only for backwards compatibility and it's recommended to switch to an option named "usethis.description".

Defaults built into usethis.

The fields discovered via options or the usethis package can be viewed with use_description_defaults().

If you create a lot of packages, consider storing personalized defaults as a named list in an option named "usethis.description". Here's an example of code to include in .Rprofile:

options(
  usethis.description = list(
    `Authors@R` = 'person("Jane", "Doe", email = "jane@example.com", role = c("aut", "cre"),
                          comment = c(ORCID = "YOUR-ORCID-ID"))',
    License = "MIT + file LICENSE",
    Language =  "es"
  )
)

use_description(fields = NULL)

use_description_defaults()

Arguments

fields
A named list of fields to add to DESCRIPTION, potentially overriding default values. See use_description() for how you can set personalized defaults using package options

See also

Examples

# NOT RUN {
use_description()

use_description(fields = list(Language = "es"))

use_description_defaults()
# }