Manage GitHub issue labels

use_github_labels() can create new labels, update colours and descriptions, and optionally delete GitHub's default labels (if delete_default = TRUE). It will never delete labels that have associated issues.

use_tidy_labels() calls use_github_labels() with tidyverse conventions powered by tidy_labels(), tidy_labels_rename(), tidy_label_colours() and tidy_label_descriptions().

use_github_labels(
  repo_spec = deprecated(),
  labels = character(),
  rename = character(),
  colours = character(),
  descriptions = character(),
  delete_default = FALSE,
  host = deprecated(),
  auth_token = deprecated()
)

use_tidy_labels(
  repo_spec = deprecated(),
  host = deprecated(),
  auth_token = deprecated()
)

tidy_labels()

tidy_labels_rename()

tidy_label_colours()

tidy_label_descriptions()

Arguments

repo_spec, host, auth_token
: These arguments are now deprecated and will be removed in the future. Any input provided via these arguments is not used. The target repo, host, and auth token are all now determined from the current project's Git remotes.

labels
A character vector giving labels to add.

rename
A named vector with names giving old names and values giving new names.

colours, descriptions
Named character vectors giving hexadecimal colours (like e02a2a) and longer descriptions. The names should match label names, and anything unmatched will be left unchanged. If you create a new label, and don't supply colours, it will be given a random colour.

delete_default
If TRUE, removes GitHub default labels that do not appear in the labels vector and that do not have associated issues.

Label usage

Labels are used as part of the issue-triage process, designed to minimise the time spent re-reading issues. The absence of a label indicates that an issue is new, and has yet to be triaged.

reprex indicates that an issue does not have a minimal reproducible example, and that a reply has been sent requesting one from the user.

bug indicates an unexpected problem or unintended behavior.

feature indicates a feature request or enhancement.

docs indicates an issue with the documentation.

wip indicates that someone is working on it or has promised to.

good first issue indicates a good issue for first-time contributors.

help wanted indicates that a maintainer wants help on an issue.

Examples

if (FALSE) {
# typical use in, e.g., a new tidyverse project
use_github_labels(delete_default = TRUE)

# create labels without changing colours/descriptions
use_github_labels(
  labels = c("foofy", "foofier", "foofiest"),
  colours = NULL,
  descriptions = NULL
)

# change descriptions without changing names/colours
use_github_labels(
  labels = NULL,
  colours = NULL,
  descriptions = c("foofiest" = "the foofiest issue you ever saw")
)
}