use_github_release {usethis}	R Documentation

Draft a GitHub release

Description

Creates a draft GitHub release for the current package using the current version and NEWS.md. If you are comfortable that it is correct, you will need to publish the release from GitHub. It also deletes CRAN-RELEASE and checks that you've pushed all commits to GitHub.

Usage

use_github_release(host = deprecated(), auth_token = deprecated())

Arguments

host, auth_token
: No longer consulted now that usethis allows the gh package to lookup a token based on a URL determined from the current project's GitHub remotes.

[Package usethis version 2.0.1 Index]