use_pkgdown {usethis}	R Documentation

Use pkgdown

Description

pkgdown makes it easy to turn your package into a beautiful website. There are two helper functions:

use_pkgdown(): creates a pkgdown config file, adds relevant files or directories to .Rbuildignore and .gitignore, and builds favicons if your package has a logo

use_pkgdown_travis(): helps you set up pkgdown for automatic deployment on Travis-CI. As part of a general pivot away from Travis-CI and towards GitHub Actions, the tidyverse team now builds pkgdown sites with a workflow configured via use_github_action("pkgdown"). use_pkgdown_travis() is still here, but is no longer actively exercised by us. Key steps:

Creates an empty gh-pages branch for the pkgdown site to be deployed to.

Prompts you about what to do next regarding Travis-CI deployment keys and updating your .travis.yml.

Usage

use_pkgdown(config_file = "_pkgdown.yml", destdir = "docs")

use_pkgdown_travis()

Arguments

config_file
Path to the pkgdown yaml config file

destdir
Target directory for pkgdown docs

See Also

https://pkgdown.r-lib.org/articles/pkgdown.html#configuration

[Package usethis version 1.6.1 Index]