use_r {usethis}	R Documentation

Create or edit a .R file

Description

Create or edit a .R file

Usage

use_r(name = NULL)

Arguments

name
File name, without extension; will create if it doesn't already exist. If not specified, and you're currently in a test file, will guess name based on test name.

See Also

use_test(), and also the R code chapter of R Packages.

[Package usethis version 1.5.0 Index]