usethis-defunct {usethis}	R Documentation

Defunct and deprecated functions in usethis

Description

These functions have either been deprecated or removed from usethis.

Usage

pr_pull_upstream()

pr_sync()

browse_github_token(...)

browse_github_pat(...)

github_token()

git_branch_default()

use_tidy_labels()

use_tidy_ci(...)

use_github_action_check_full(
  save_as = "R-CMD-check.yaml",
  ignore = TRUE,
  open = FALSE,
  repo_spec = NULL
)

pr_pull_upstream()

This function has been replaced by pr_merge_main().

pr_sync()

Bundling these operations together did not seem justified, in terms of how rarely this comes up and, when it does, how likely merge conflicts are. Users of pr_sync() should implement these steps "by hand":

(Check you are on a PR branch)

pr_pull()

pr_merge_main(), deal with any merge conflicts, if any

pr_push()

browse_github_token(), browse_github_pat()

These functions have been replaced by create_github_token().

github_token()

All implicit and explicit token discovery routes through gh::gh_token() now.

git_branch_default()

Please call git_default_branch() instead. In hindsight, that is a better name for this function.

use_tidy_labels()

Please call use_tidy_github_labels() instead. In hindsight, that is a better name for this function.

[Package usethis version 2.1.5 Index]