usethis-package {usethis}	R Documentation

usethis: Automate Package and Project Setup

Description

Automate package and project setup tasks that are otherwise performed manually. This includes setting up unit testing, test coverage, continuous integration, Git, 'GitHub', licenses, 'Rcpp', 'RStudio' projects, and more.

Author(s)

Maintainer: Jennifer Bryan jenny@posit.co (ORCID)

Authors:

Hadley Wickham hadley@posit.co (ORCID)

Malcolm Barrett malcolmbarrett@gmail.com (ORCID)

Andy Teucher andy.teucher@posit.co (ORCID)

Other contributors:

Posit Software, PBC [copyright holder, funder]

See Also

Useful links:

https://usethis.r-lib.org

https://github.com/r-lib/usethis

Report bugs at https://github.com/r-lib/usethis/issues

[Package usethis version 2.2.3 Index]